+++ oceantracker/particle_properties/inside_polygons.py
@@ -1,7 +1,7 @@
 from oceantracker.particle_properties._base_properties import ParticleProperty
 import numpy as np
 from oceantracker.util.parameter_checking import ParamDictValueChecker as PVC, ParameterListChecker as PLC
-from oceantracker.util.polygon_util import set_up_list_of_polygon_instances
+from oceantracker.util.polygon_util import set_up_list_of_polygon_instances, InsidePolygon
 from oceantracker.common_info_default_param_dict_templates import default_polygon_dict_params
 
 class InsidePolygonsNonOverlapping2D(ParticleProperty):
@@ -32,9 +32,6 @@ class InsidePolygonsNonOverlapping2D(ParticleProperty):
         self.polygons, msg = set_up_list_of_polygon_instances(self.params['polygon_list'])
         si.case_log.add_messages(msg)
 
-    def initial_value_at_birth(self, new_part_IDs):
-        self.set_values(self.params['initial_value'], new_part_IDs) # sets this properties values
-
     def update(self, active):
         # find polygon each particle is inside
         part_prop = self.shared_info.classes['particle_properties']
@@ -52,3 +49,4 @@ class InsidePolygonsNonOverlapping2D(ParticleProperty):
                                          also_return_indices_outside=True, out_outside=self.get_particle_subset_buffer())
             self.set_values(n, inside)
             to_search = out_side # on next polygon only search those not in this polygon
+