+++ oceantracker/shared_info.py
@@ -9,27 +9,30 @@ class SharedInfoClass(object):
 
     def reset(self):
         self.classes = {}
-        self.class_list_interators = {}
+        self.class_interators_using_name = {}
         self.core_class_interator = {}
+        self.classes_as_lists ={}
         # fill in known user class and iterator names
         for key, item in default_case_param_template.items():
             if type(item) == list:
                 self.classes[key] = {}
-                self.class_list_interators[key] = {'all': {}, 'user': {} ,'manual_update':{}}
+                self.classes_as_lists[key] = []
+                self.class_interators_using_name[key] = {'all': {}, 'user': {} , 'manual_update':{}}
+
 
     def add_core_class(self,class_type,class_params,  make_core=False):
         cl= self.case_log
         if class_type not in default_case_param_template and not make_core:
             cl.write_msg('add_core_class, name is not a known core class, name=' + class_type,
                          crumbs='Adding core class type=' + class_type,
-                         exception = GracefulExitError, raiseerrors=True)
+                         exception = GracefulExitError)
 
         i, msg = import_module_from_string(class_params['class_name'].strip())
-        cl.add_messages(msg, raiseerrors=True)
+        cl.add_messages(msg)
 
         # merge params
         msg_list = i.merge_with_class_defaults(class_params, {}, crumbs='Merging core classes with defaults >>> ' + class_type )
-        self.case_log.add_messages(msg_list, raiseerrors=True)
+        self.case_log.add_messages(msg_list)
 
         self.classes[class_type]= i
         self.core_class_interator[class_type] = i
@@ -37,11 +40,11 @@ class SharedInfoClass(object):
 
     def create_class_interator(self,class_type, known_iteration_groups=None):
         if class_type not in self.classes: self.classes[class_type] = {}
-        if class_type not in self.class_list_interators: self.class_list_interators[class_type] = {'all': {}}
+        if class_type not in self.class_interators_using_name: self.class_interators_using_name[class_type] = {'all': {}}
 
         if known_iteration_groups is not None:
             for g in known_iteration_groups:
-                self.class_list_interators[class_type].update({g :{}})
+                self.class_interators_using_name[class_type].update({g :{}})
 
     def add_class_instance_to_list_and_merge_params(self, class_type, iteration_group, class_params, crumbs=''):
         # dynamically  get instance of class from string eg oceantracker.solver.Solver
@@ -54,27 +57,27 @@ class SharedInfoClass(object):
                 known_types.append(key)
 
         if class_type not in known_types:
-            cl.write_msg('add_to_class_list: name is not a known class list,class_type=' + class_type , exception = GracefulExitError, crumbs = crumbs, raiseerrors=True)
+            cl.write_msg('add_to_class_list: name is not a known class list,class_type=' + class_type , exception = GracefulExitError, crumbs = crumbs)
 
         if 'class_name' not in class_params and class_type in default_class_names:
             class_params['class_name']= default_class_names[class_type]
 
         i, msg = import_module_from_string(class_params['class_name'])
-        cl.write_msg(msg, raiseerrors=True, crumbs= 'Importing class >>> '+  crumbs)
+        cl.write_msg(msg,  crumbs= 'Importing class >>> '+  crumbs)
 
-        i.info['instanceID'] = len(self.class_list_interators[class_type][iteration_group])
+        i.info['instanceID'] = len(self.class_interators_using_name[class_type][iteration_group])
         nseq = i.info['instanceID']  + 1
 
         # merge to get any default class name and params
         msg_list = i.merge_with_class_defaults(class_params, {}, crumbs = crumbs+ ' >>> Merging with class defaults >>> ' + class_type + '[#' + str(nseq) + '] ')
-        self.case_log.add_messages(msg_list, raiseerrors=True)
+        self.case_log.add_messages(msg_list)
 
         if i.params['name'] is None or i.params['name'] =='':
             if iteration_group == 'user':
-                i.params['name'] = 'unnamed%03.0f' %  (len(self.class_list_interators[class_type]['user'])+1)
+                i.params['name'] = 'unnamed%03.0f' %  (len(self.class_interators_using_name[class_type]['user']) + 1)
             else:
                 # this may be redundent if name param is required by class
-                cl.write_msg('Only user added classes can be unnamed, all others must must have param["name"]' , exception = GracefulExitError, raiseerrors=True,
+                cl.write_msg('Only user added classes can be unnamed, all others must must have param["name"]' , exception = GracefulExitError,
                              crumbs= crumbs + ' >>> ' + class_params['class_name'] )
 
         name = i.params['name']
@@ -82,7 +85,7 @@ class SharedInfoClass(object):
         if name in self.classes[class_type]:
             cl.write_msg('Class type"' + class_type + '" already has a class with name = "' + i.params['name']
                          + '", "name" parameter must be unique',
-                         crumbs = ' Checking for unique class names >>> '+  crumbs, exception = GracefulExitError, raiseerrors=True)
+                         crumbs = ' Checking for unique class names >>> '+  crumbs, exception = GracefulExitError)
 
         # check class type OK
         known_types= []
@@ -95,9 +98,9 @@ class SharedInfoClass(object):
 
         # now add to class lists and interators
 
-        if iteration_group not in self.class_list_interators[class_type]:
+        if iteration_group not in self.class_interators_using_name[class_type]:
             cl.write_msg('add_to_class_list: iteration_group  for class_type=' + class_type + ', group="'
-                                    + iteration_group + '", is not one of known types=' + str(self.class_list_interators[class_type].keys()), exception = GracefulExitError)
+                         + iteration_group + '", is not one of known types=' + str(self.class_interators_using_name[class_type].keys()), exception = GracefulExitError)
 
         return i
 
@@ -105,8 +108,11 @@ class SharedInfoClass(object):
     def add_class_instance_to_interators(self, name, class_type, iteration_group, i):
         i.info['instanceID'] = len(self.classes[class_type]) # needed for release group identification info etc
         self.classes[class_type][name] = i
-        self.class_list_interators[class_type]['all'][name] = i
-        self.class_list_interators[class_type][iteration_group][name] = i
+        self.classes_as_lists[class_type].append(i)
+
+        self.class_interators_using_name[class_type]['all'][name] = i
+        self.class_interators_using_name[class_type][iteration_group][name] = i
+